restructure `ott-vis` to allow for multiple visualization types
