+++ packages/ott-vis/src/module.ts
@@ -1,9 +1,27 @@
 import { PanelPlugin } from "@grafana/data";
-import { SimpleOptions } from "./types";
-import { SimplePanel } from "./components/SimplePanel";
+import { CoreOptions } from "./types";
+import { CorePanel } from "./components/CorePanel";
 
-export const plugin = new PanelPlugin<SimpleOptions>(SimplePanel).setPanelOptions(builder => {
+export const plugin = new PanelPlugin<CoreOptions>(CorePanel).setPanelOptions(builder => {
 	return builder
+		.addSelect({
+			path: "view",
+			name: "View",
+			description: "Select the view to display",
+			defaultValue: "global",
+			settings: {
+				options: [
+					{
+						value: "global",
+						label: "Global",
+					},
+					{
+						value: "region",
+						label: "Region",
+					}
+				],
+			},
+		})
 		.addTextInput({
 			path: "text",
 			name: "Allowed Entities",
