+++ packages/ott-vis/src/types.ts
@@ -1,7 +1,30 @@
 type SeriesSize = "sm" | "md" | "lg";
 
-export interface SimpleOptions {
+export interface CoreOptions {
+	view: "global" | "region";
 	text: string;
 	showSeriesCount: boolean;
 	seriesCountSize: SeriesSize;
+}
+
+/**
+ * Describes the state of OTT
+ */
+export type SystemState = Balancer[]
+
+export interface Balancer {
+	id: string;
+	region: string;
+	monoliths: Monolith[];
+}
+
+export interface Monolith {
+	id: string;
+	region: string;
+	rooms: Room[];
+}
+
+export interface Room {
+	name: string;
+	clients: number;
 }
